packages/service-main-py/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/service_main"]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "B"]
ignore = []

[tool.ruff.format]
quote-style = "double"

[project]
name = "service-main"
description = ""
version = "0.0.1"
license = "MIT"
authors = [{ name = "YG Min", email = "ygmindev@gmail.com" }]
requires-python = "==3.13.12"
dependencies = [
    "beanie>=2.1.0",
    "lib-ai",
    "lib-config",
    "lib-model",
    "lib-shared",
    "pydash>=8.0.6",
    "strawberry-graphql>=0.315.1",
]

[tool.uv.sources]
lib-ai = { path = "../lib-ai-py", editable = true }
lib-config = { path = "../lib-config-py", editable = true }
lib-model = { path = "../lib-model-py", editable = true }
lib-shared = { path = "../lib-shared-py", editable = true }

[tool.mypy]
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_defs = true
ignore_missing_imports = true
warn_unused_configs = true
